esp32: do SoC reset before resetting the CPU
If peripherals are in use, resetting the CPU may not be sufficient to
achieve a clean reset. Peripherals may still be using DMA, which will
keep working after CPU reset, and may corrupt memory and cause other
side effects. This change implements almost full SoC reset, except for
the state of RTC_CNTL which can not be reset automatically. Relevant
registers of RTC_CNTL are restored to the default state explicitly.