esp32: go directly into halted state in xtensa_step
Previous behaviour of stepping in Xtensa target was different from
most other OpenOCD targets. The target was kept in running state, and
xtensa_poll function would handle the fact that CPU is halted and
return control to the debugger. This caused stepping to be slow in
many cases, since xtensa_poll is only called once every 100ms. This
changes xtensa_step to indicate that target is halted. This allows
gdb to do the next step immediately instead of waiting for the next
polling period.

Related to https://github.com/espressif/openocd-esp32/issues/29