esp32: skip window over/underflow exceptions when stepping
If isrmask option is enabled, stepping disables interrupts from
disrupting the debugging flow. Window exceptions are not masked by
PS.INTLEVEL, and can’t be skipped the same way as high level
interrupts, via ICOUNTLEVEL. This change adds code to detect whether
execution is happening inside window exception and to step out of it.

This also happens to work around GDB assert when reading EPC1, since
that happens in the code path for window exceptions.

Closes https://github.com/espressif/openocd-esp32/issues/37