Don't update RTOS threads just before step. (#376)
This messes up all kinds of tests against HiFive Unleashed, because some
harts may be single stepped from previous tests. The symptom is that gdb
will suddenly be accessing a different hart than you think it is.

I replaced it with a comment so I can remember what happened when time
comes to upstream this change. It may not be acceptable depending on
what the reason for the call is in the first place.

Change-Id: I1fb44d5a7792835f66342f590a5f7bbf8c21b64e