contrib/firmware/angie: reorganize the endpoints of the microcontroller
The new firmware of ANGIE will not use Bitbang method to
transfer jtag data to the target chip. instead, it will use the
the GPIF module to bypass JTAG data directly to the FPGA and
then to target chip.

So we delete the protocol and jtag files which handle bitbang.

We are going to use endpoint 2/4 for OUT/IN GPIF transactions,
and we deactivate the endpoints 1IN and 1OUT.

we will keep the endpoint 6/8 for i2c unchanged.

Change-Id: I0fcb23690526f6a7da044b702217b32522be727a
Signed-off-by: Ahmed BOUDJELIDA <aboudjelida@nanoxplore.com>
Reviewed-on: https://review.openocd.org/c/openocd/+/8712
Tested-by: jenkins
Reviewed-by: Antonio Borneo <borneo.antonio@gmail.com>