replacements: make socket_select() work on Windows if no sockets are provided
On Windows 11, if select() is called with empty sets, it fails and
returns WSAINVAL. On POSIX this works fine.

This patch addresses it by detecting this case in OpenOCD replacements
and returning 0 in these cases.

This fixes OpenOCD crash on Windows if no services are enabled
(gdb server, tcl server and telnet server all disabled).

Change-Id: I601878671caf4ae44e105d6a819251d2d96c607c
Signed-off-by: Marek Vrbka <marek.vrbka@codasip.com>
Reviewed-on: https://review.openocd.org/c/openocd/+/9081
Tested-by: jenkins
Reviewed-by: Antonio Borneo <borneo.antonio@gmail.com>